Lots of changes, reflection partially working!
Made a lot of changes, fixed and added various issues. The code is a mess and a large percentage of it are specific collision functions. This desperately needs refactoring, including splitting into multiple files, which is what I'll do soon. Reflection mostly works, no distance comparison yet, made the ray star move with the cursor and added an option to set the amount of spokes. Visual change where segments that go off into infinity are rendered lighter than the ones that end up colliding with something